# Improve EqualRowNumber